+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeDocumentsEndpoint.restore`
 - Add `CompletedIdentificationReport.name`
 - Add `ReportEndpoint.completedIdentifications` endpoint and `CompletedIdentificationReport` resource
+- Add `IdentificationResult` claims
 
 ## [2.3.0] - 2024-03-28
 ### Added